# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 Notebook OpenApi",
        "description": "L5 Swagger OpenApi for Notebook",
        "contact": {
            "email": "wolk0193@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/notebooks/{id}": {
            "get": {
                "tags": [
                    "Notebooks"
                ],
                "description": "Home page",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Notebooks"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "email"
                                            },
                                            "birth_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "photo": {
                                                "type": "string"
                                            },
                                            "company": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "phone": "836518294",
                                    "email": "test@mail.test",
                                    "birthDate": "1999-11-11 11:11:11",
                                    "photo": "photo.jpg",
                                    "company": "company name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebooks"
                ],
                "description": "Home page",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Notebooks"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "email"
                                            },
                                            "birth_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "photo": {
                                                "type": "string"
                                            },
                                            "company": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "email": "test@mail.test",
                                    "company": "company name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/notebooks": {
            "get": {
                "tags": [
                    "Notebooks"
                ],
                "description": "Home page",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebooks"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "email"
                                            },
                                            "birth_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "photo": {
                                                "type": "string"
                                            },
                                            "company": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "phone": "836518294",
                                    "email": "test@mail.test",
                                    "birthDate": "1999-11-11 11:11:11",
                                    "photo": "photo.jpg",
                                    "company": "company name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/notebooks/bulk": {
            "post": {
                "tags": [
                    "Notebooks"
                ],
                "description": "Use an array of objects to insert several items at once",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "email"
                                            },
                                            "birth_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "photo": {
                                                "type": "string"
                                            },
                                            "company": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "phone": "836518294",
                                    "email": "test@mail.test",
                                    "birthDate": "1999-11-11 11:11:11",
                                    "photo": "photo.jpg",
                                    "company": "company name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}